Use user filtered relevant tables in MB state (#104)
