+++ apps/src/package.ts
@@ -1,1 +1,2 @@
 export { getAppStateConfigs } from "./appStateConfigs";
+export { applyTableDiffs } from "./common/utils";
