+++ web/src/types.d.ts
@@ -2,3 +2,4 @@ export type { DefaultMessageContent, BlankMessageContent, ActionRenderInfo } fro
 export type { GoogleState } from './state/google/types';
 export type { SemanticFilter, TimeDimension, Order, SemanticQuery } from './state/thumbnails/reducer';
 export type { Measure, Dimension } from './state/semantic-layer/reducer';
+export type { TableDiff } from './state/settings/reducer';
